==> scripts/python/view_experimentation/2_3_gmres_m_comparison.ipynb <==
import os
import matplotlib
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import matplotlib.gridspec
from typing import List, Tuple

from exp_spec_info import *
from plot_info import *
from select_data import *

# Processed median path
median_data_path = "C:\\Users\\dosre\\dev\\thesis-data\\median_data.pkl"

# Matrix data
small_matrix_data_path = "C:\\Users\\dosre\\dev\\precision-cascade\\scripts\\python\\view_experimentation\\data\\small-experimental-matrix-data.csv"
large_matrix_data_path = "C:\\Users\\dosre\\dev\\precision-cascade\\scripts\\python\\view_experimentation\\data\\large-experimental-matrix-data.csv"

# Plot output directory
chap_6_image_dir = "C:\\Users\\dosre\\Desktop\\MSCthesis\\thesis\\body_chapters\\chap_6\\images"
os.makedirs(chap_6_image_dir, exist_ok=True)
unprecond_dir = os.path.join(chap_6_image_dir, "unprecond-analysis")
os.makedirs(unprecond_dir, exist_ok=True)
precond_dir = os.path.join(chap_6_image_dir, "precond-analysis")
os.makedirs(precond_dir, exist_ok=True)
chap_a3_image_dir = "C:\\Users\\dosre\\Desktop\\MSCthesis\\thesis\\body_chapters\\chap_a3\\images"
os.makedirs(chap_a3_image_dir, exist_ok=True)
chap_a4_image_dir = "C:\\Users\\dosre\\Desktop\\MSCthesis\\thesis\\body_chapters\\chap_a4\\images"
os.makedirs(chap_a4_image_dir, exist_ok=True)

# Load matrix data
small_matrix_data = pd.read_csv(small_matrix_data_path).rename({"name": "matrix"}, axis=1)
large_matrix_data = pd.read_csv(large_matrix_data_path).rename({"name": "matrix"}, axis=1)
all_matrix_data = pd.concat([small_matrix_data.drop("cond", axis=1), large_matrix_data])

# Load data and filter for only gmres m solvers
median_data = pd.read_pickle(median_data_path)
unprecond_median_data = median_data[
    (median_data["solver"].transform(lambda x: (x in PC_SOLVERS))) &
    median_data["setup"].isin({"unprecond", "unpreconddense"})
]
precond_median_data = median_data[
    (median_data["solver"].transform(lambda x: (x in PC_SOLVERS))) &
    median_data["setup"].isin({"ilu0", "ilutp1em2", "ilutp1em4"})
]

##### 6.3/6.4 Prune Non-Complaint Experiments

fractional_error = 0.10

def prune_analyze_non_compliant_relres_data(
    data: pd.DataFrame, threshold: float, plot_title: str, plot_save_path: str
) -> pd.DataFrame:

    failed_fp64_exp = (
        (data["med_rel_res"]/(data["med_rel_res_frac_err_fp64"] + 1)) > 1
    )
    keep = ~failed_fp64_exp & (data["med_rel_res_frac_err_fp64"] <= threshold)
    compliant_data = data[keep]
    eliminated_data_for_analysis = data[(~failed_fp64_exp & (data["med_rel_res_frac_err_fp64"] > threshold))]

    elim_mat = eliminated_data_for_analysis["matrix"].unique()
    elim_mat = elim_mat[
        np.argsort([mat.lower() for mat in elim_mat])
    ]
    n_elim_mat = len(elim_mat)
    mat_index_dict = dict(zip(elim_mat, range(n_elim_mat)))
    index_mat_dict = dict(zip(range(n_elim_mat), elim_mat))

    solver_percents = eliminated_data_for_analysis["solver"].value_counts()/len(eliminated_data_for_analysis)
    for solver in PC_SOLVERS:
        if solver not in solver_percents:
            solver_percents[solver] = 0
    restart_param_percents = eliminated_data_for_analysis["restart_param"].value_counts()/len(eliminated_data_for_analysis)
    for restart_param in RESTART_PARAMS:
        if restart_param not in restart_param_percents:
            restart_param_percents[restart_param] = 0

    fig = plt.figure(figsize=(6.4, 5.6))
    spec = matplotlib.gridspec.GridSpec(
        ncols=2,
        nrows=3,
        width_ratios=[0.6, 0.4],
        height_ratios=[0.36, 0.38, 0.25],
        hspace=0.05,
        wspace=0.05,
        figure=fig
    )
    ax = fig.add_subplot(spec[:, 0])
    ax1 = fig.add_subplot(spec[0, 1])
    ax2 = fig.add_subplot(spec[1, 1])
    ax3 = fig.add_subplot(spec[2, 1])

    for solver in PC_SOLVERS:
        select = eliminated_data_for_analysis["solver"] == solver
        ax.scatter(
            eliminated_data_for_analysis["med_rel_res_frac_err_fp64"][select],
            eliminated_data_for_analysis["matrix"][select].transform(
                lambda mat: mat_index_dict[mat]
            ),
            eliminated_data_for_analysis["restart_param"][select],
            marker=".",
            color=SOLVER_CLR_DICT[solver],
            label=solver
        )

    ax.grid()
    ax.set_axisbelow(True)

    ax1.legend(
        handles=[matplotlib.lines.Line2D([0], [0], marker="o", linestyle="None", color=SOLVER_CLR_DICT[solver]) for solver in PC_SOLVERS],
        labels=[f"{solver} ({solver_percents[solver]:.2f})" for solver in PC_SOLVERS],
        loc="upper left",
        bbox_to_anchor = (0., 1.),
        borderaxespad=0.
    )
    ax2.legend(
        handles=[matplotlib.lines.Line2D([0], [0], marker=".", markersize=np.sqrt(val), linestyle="None", color="grey") for val in RESTART_PARAMS],
        labels=[f"$m$={val} ({restart_param_percents[val]:.2f})" for val in RESTART_PARAMS],
        loc="upper left",
        bbox_to_anchor = (0., 1.),
        borderaxespad=0.
    )
    
    elim_data_frac = len(eliminated_data_for_analysis)/len(data)
    table = ax3.table(
        [
            ["FP64 Failed Frac."],
            [f"{failed_fp64_exp.sum()/len(data):.2g}"],
            ["Non-Compliant Frac."],
            [f"{elim_data_frac:.2g}"],
        ],
        bbox=(0., 0., .9, .7)
    )
    table.auto_set_font_size(False)
    for info_ax in [ax1, ax2, ax3]:
        info_ax.tick_params(
            axis="both", bottom=False, left=False, labelbottom=False, labelleft=False
        )
        info_ax.axis("off")

    ax.semilogx()
    ax.set_xlabel("Med. Rel. Res. Frac. Error to FP64")
    ax.set_yticks(range(n_elim_mat))
    ax.set_yticklabels([index_mat_dict[i] for i in range(n_elim_mat)])
    ax.set_ylabel("Linear Solve Matrix")
    ax.set_title(plot_title)

    plt.savefig(plot_save_path, bbox_inches="tight")
    plt.show()
    plt.close()

    return compliant_data

unprecond_compliant_data = prune_analyze_non_compliant_relres_data(
    unprecond_median_data,
    fractional_error,
    f"Non-Compliant Unpreconditioned Experiments to Relative Residual\n Fractional Error Threshold of {fractional_error:.2f}",
    os.path.join(unprecond_dir, "unprecond-threshold-non-compliance.png")
)

precond_compliant_data = prune_analyze_non_compliant_relres_data(
    precond_median_data,
    fractional_error,
    f"Non-Compliant Preconditioned Experiments to Relative Residual\n Fractional Error Threshold of {fractional_error:.2f}",
    os.path.join(precond_dir, "precond-threshold-non-compliance.png")
)

for solver in PC_SOLVERS:
    n_compliant = len(unprecond_compliant_data[unprecond_compliant_data["solver"] == solver])
    n_total = len(unprecond_median_data[unprecond_median_data["solver"] == solver])
    print(f"{solver}: {n_compliant/n_total:.2f}")

(0.83+0.89+0.86+0.91)/4-(0.97+0.97+0.98)/3

fp_data = median_data[
    median_data["solver"].isin({"FP FP16", "FP FP32"}) &
    median_data["setup"].isin({"unprecond", "unpreconddense"})
]
failed_fp64_exp = (
    (fp_data["med_rel_res"]/(fp_data["med_rel_res_frac_err_fp64"] + 1)) > 1
)
keep = ~failed_fp64_exp & (fp_data["med_rel_res_frac_err_fp64"] <= fractional_error)
compliant_fp_data = fp_data[keep]
fp_data = fp_data[~failed_fp64_exp]

for solver in ["FP FP16", "FP FP32"]:
    n_compliant = len(compliant_fp_data[compliant_fp_data["solver"] == solver])
    n_total = len(fp_data[fp_data["solver"] == solver])
    print(f"{solver}: {n_compliant/n_total:.2f}")

failed_fp64_exp = (
    (precond_median_data["med_rel_res"]/(precond_median_data["med_rel_res_frac_err_fp64"] + 1)) > 1
)
for solver in PC_SOLVERS:
    n_compliant = len(precond_compliant_data[precond_compliant_data["solver"] == solver])
    n_total = len(precond_median_data[~failed_fp64_exp & (precond_median_data["solver"] == solver)])
    print(f"{solver}: {n_compliant/n_total:.2f}")

fp_data = median_data[
    median_data["solver"].isin({"FP FP16", "FP FP32"}) &
    median_data["setup"].isin({"ilu0", "ilutp1em2", "ilutp1em4"})
]
failed_fp64_exp = (
    (fp_data["med_rel_res"]/(fp_data["med_rel_res_frac_err_fp64"] + 1)) > 1
)
keep = ~failed_fp64_exp & (fp_data["med_rel_res_frac_err_fp64"] <= fractional_error)
compliant_fp_data = fp_data[keep]
fp_data = fp_data[~failed_fp64_exp]

for solver in ["FP FP16", "FP FP32"]:
    n_compliant = len(compliant_fp_data[compliant_fp_data["solver"] == solver])
    n_total = len(fp_data[fp_data["solver"] == solver])
    print(f"{solver}: {n_compliant/n_total:.2f}")

precond_compliant_data[
    # precond_compliant_data["all_converged"] &
    (precond_compliant_data["solver"].transform(lambda x: (not "HSD" in x) and (not "HSD RRT" in x)))
]["med_inner_iter"].median()

precond_compliant_data[
    # precond_compliant_data["all_converged"] &
    (precond_compliant_data["solver"].transform(lambda x: ("HSD" in x) and (not "HSD RRT" in x)))
]["med_inner_iter"].median()

##### 6.3/6.4 Unpreconditioned FP64 FP-GMRES($m$) and PC-GMRES($m$) Relative Time Comparison

def analyze_rel_time(
    data: pd.DataFrame,
    setups: List[str],
    setup_formats: List[Tuple[str, str]],
    x_range: Tuple[float, float],
    plot_title_prefix: str,
    plot_dir: str,
    plot_save_file_title_prefix: str,
):

    matrices = data["matrix"].unique()
    n_matrices = len(matrices)
    mat_index_dict = dict(zip(matrices, range(n_matrices)))
    index_mat_dict = dict(zip(range(n_matrices), matrices))

    for restart_param in RESTART_PARAMS:

        restart_data = data[data["restart_param"] == restart_param]

        fig, axs = plt.subplots(
            1, len(PC_SOLVERS), figsize=(11, 8), sharey=True
        )

        for ax, solver in zip(axs, PC_SOLVERS):

            solver_data = restart_data[restart_data["solver"] == solver]

            for setup, setup_format in zip(setups, setup_formats):

                marker, line = setup_format

                current_data = solver_data[solver_data["setup"] == setup]

                ax.scatter(
                    current_data["med_rel_time_fp64"],
                    current_data["matrix"].transform(lambda mat: mat_index_dict[mat]),
                    color=SOLVER_CLR_DICT[solver],
                    marker=marker,
                    zorder=2
                )
                ax.axvline(
                    current_data["med_rel_time_fp64"].median(),
                    color=SOLVER_CLR_DICT[solver],
                    linestyle=line,
                    zorder=2
                )
                ax.axvline(1., color=SOLVER_CLR_DICT["FP FP64"], zorder=1)
                ax.set_title(f"{solver}")

                ax.set_xlim(x_range[0], x_range[1])

            rel_solve_data = solver_data[solver_data["setup"].isin(setups)]
            off_screen_data = (
                rel_solve_data[rel_solve_data["med_rel_time_fp64"] >= x_range[1]]["matrix"]
            ).value_counts()
            for matrix in off_screen_data.index:
                ax.text(
                    x_range[0]+0.99*(x_range[1]-x_range[0]),
                    mat_index_dict[matrix],
                    f"{off_screen_data.loc[matrix]}$\\rightarrow$",
                    verticalalignment='center',
                    horizontalalignment='right'
                )

        for ax in axs:
            ax.grid()
        axs[3].set_xlabel(f"Median Time Relative to FP64")
        axs[0].set_ylim(-0.75, n_matrices-0.25)
        axs[0].set_yticks(range(n_matrices))
        axs[0].set_yticklabels([index_mat_dict[i] for i in range(n_matrices)])
        axs[0].set_ylabel("Linear Solve Matrix")

        fig.suptitle(plot_title_prefix + f"PC-GMRES(${restart_param}$) Time Relative to FP64")
        fig.tight_layout()

        plt.savefig(
            os.path.join(
                plot_dir,
                f"{plot_save_file_title_prefix}-{restart_param:03d}.png"
            )
        )
        plt.show()
        plt.close()

analyze_rel_time(
    unprecond_compliant_data, ["unprecond", "unpreconddense"],
    [(".", ":"), ("x", "--")],
    (0., 3.),
    "Unpreconditioned ", chap_a3_image_dir, "unprecond-rel-time"
)

analyze_rel_time(
    precond_compliant_data, ["ilu0", "ilutp1em2", "ilutp1em4"],
    [(".", ":"), ("+", "--"), ("x", "-.")],
    (0., 20.),
    "Preconditioned ", chap_a4_image_dir, "precond-rel-time"
)


filt = unprecond_compliant_data[
    (unprecond_compliant_data["restart_param"] == 200) &
    (unprecond_compliant_data["med_rel_time_fp64"] >= 3.) &
    (unprecond_compliant_data["solver"] == "PC HSD ORC")
]

#### 6.3/6.4 PC-GMRES($m$) Comparison Matrix

comp_data = pd.concat([unprecond_compliant_data, precond_compliant_data], axis=0)
comp_data = comp_data[["setup", "restart_param", "solver", "med_rel_time_fp64"]]
comp_data = comp_data.groupby(by=["setup", "restart_param", "solver"], as_index=False).median()
comp_data = comp_data.pivot(index=["setup", "solver"], columns="restart_param", values="med_rel_time_fp64")
comp_data = comp_data.round(2)

best_rel_time = comp_data.min().min()

def pc_gmres_comp_matrix(data: pd.DataFrame, setup: str, cmap_range: Tuple[float, float], plot_save_title: str):

    setup_data = data.loc[setup].loc[PC_SOLVERS[::-1]]
    fig, ax = plt.subplots(figsize=(6.4, 4.8))

    solvers = setup_data.index.to_numpy()
    restart_params = setup_data.columns.to_numpy()

    X, Y = np.meshgrid(np.arange(0, len(restart_params)), np.arange(0, len(solvers)))
    quad_mesh = ax.pcolormesh(X, Y, setup_data.to_numpy(), cmap="Blues_r", vmin=cmap_range[0], vmax=cmap_range[1])
    for i in np.arange(0, len(restart_params)):
        for j in np.arange(0, len(solvers)):
            ax.text(i, j, setup_data.loc[solvers[j]][restart_params[i]], ha="center", va="center", fontsize=9)

    ax.set_xticks(np.arange(0, len(restart_params)))
    ax.set_xticklabels([restart_params[i] for i in np.arange(0, len(restart_params))])
    ax.set_xlabel("Restart Parameter")
    ax.set_yticks(np.arange(0, len(solvers)))
    ax.set_yticklabels([solvers[i] for i in np.arange(0, len(solvers))])
    ax.set_ylabel("PC-GMRES($m$) Solver")
    ax.set_title(f"{SETUP_NAME_MAPPING[setup]} PC-GMRES($m$) Median Time Relative to FP64")

    cbar = plt.colorbar(quad_mesh, ax=ax)
    ticks = np.linspace(cmap_range[0], cmap_range[1], 5)
    cbar.ax.set_yticks(ticks)
    cbar.ax.set_yticklabels(
        [f"{tick:.2f}" for tick in ticks[:len(ticks)-1]] + [f"{cmap_range[1]:.2f}+"]
    )
    plt.savefig(plot_save_title, bbox_inches="tight")
    plt.show()

for setup, plot_dir in [
    ("unprecond", unprecond_dir),
    ("unpreconddense", unprecond_dir),
    ("ilu0", precond_dir),
    ("ilutp1em2", precond_dir),
    ("ilutp1em4", precond_dir)
]:
    pc_gmres_comp_matrix(
        comp_data,
        setup,
        (best_rel_time, 1.),
        os.path.join(plot_dir, f"rel-time-summary-{setup}.png")
    )

##### Misc. Analysis

small_merge = unprecond_compliant_data.merge(small_matrix_data, on="matrix", how="inner")
small_merge = small_merge[small_merge["setup"] == "unpreconddense"]
large_merge = unprecond_compliant_data.merge(large_matrix_data, on="matrix", how="inner")
large_merge = large_merge[large_merge["setup"] == "unpreconddense"]

fig, ax = plt.subplots()
ax.scatter(small_merge["med_inner_iter"], small_merge["med_rel_time_fp64"], marker=".")
ax.scatter(large_merge["med_inner_iter"], large_merge["med_rel_time_fp64"], marker=".")
ax.semilogx()
ax.set_yscale("symlog")
print(np.corrcoef(
    np.hstack([small_merge["med_inner_iter"], large_merge["med_inner_iter"]]),
    np.hstack([small_merge["med_rel_time_fp64"], large_merge["med_rel_time_fp64"]])
))
plt.show()

fig, ax = plt.subplots()
ax.scatter(small_merge["nnz"], small_merge["med_rel_time_fp64"], marker=".")
ax.scatter(large_merge["nnz"], large_merge["med_rel_time_fp64"], marker=".")
# ax.semilogx()
# ax.set_yscale("symlog")
print(np.corrcoef(
    np.hstack([small_merge["nnz"], large_merge["nnz"]]),
    np.hstack([small_merge["med_rel_time_fp64"], large_merge["med_rel_time_fp64"]])
))
plt.show()

fig, ax = plt.subplots()
ax.scatter(small_merge["rows"], small_merge["med_rel_time_fp64"], marker=".")
ax.scatter(large_merge["rows"], large_merge["med_rel_time_fp64"], marker=".")
# ax.semilogx()
# ax.set_yscale("symlog")
print(np.corrcoef(
    np.hstack([small_merge["rows"], large_merge["rows"]]),
    np.hstack([small_merge["med_rel_time_fp64"], large_merge["med_rel_time_fp64"]])
))
plt.show()

fig, ax = plt.subplots()
ax.scatter(small_merge["cond"], small_merge["med_rel_time_fp64"], marker=".")
print(np.corrcoef(small_merge["cond"], small_merge["med_rel_time_fp64"]))
plt.show()
